requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
openai==1.35.10
python-multipart==0.0.9
Pillow==10.4.0
